Implement `test` wrapper procedure
In Ruby Minitest `test` would be implemented something like this:
```
def test(name = 'untitled')
  yield
end

// In the unit test
test('some test description here' do
  assert_equal 2, 1 + 1
end
```

A block in Ruby is an anonymous method object declared on the spot.
Object Pascal does not have such a concept. You can however provide the
address of a procedure using `@` to another procedure argument. The
address is to a procedure which serves as the 'body' of the `test`
procedure. This mimics how `test` is used in Ruby. The downside is that
in Pascal the `test` call and procedure definition are in two separate
places. It's not as easy to read as Minitest but it does work.