Remove `test` procedure and instead use test methods directly
Following the along with the YouTube video, we just use built in Pascal
procedures directly. The procedure name serves as the test name. This
keeps test state isolated to procedure scope.

Removing the use of `test` makes the implementation simpler. It is also
more readable since the reader does not have to look in one place for
the text declaration and another for the procedure definition.

This also puts us in the ideal place to refactor the test procedures
into a class.