Promote test procedures into a Test class
By collecting all the tests into a class it will make it easier to
automate running tests with future changes.

Note that in the test program heap allocation is taking place with each
`create` call but it is not being explicitly freed. This is OK for now
as the program is small and will rapidly change.