Remove redundant Makefile targets
It looks like `fpc` is smart enough to find and compile and dependency
trees e.g. `TestProgram.pas` depends on `Test.pas` which depends on
`PascalTest.pas`. Therefore the Makefile does not need to have rules for
these dependent units, making it simpler.

I'm not 100% sure but doing it this way seems to create *.res compiled
resource files as a byproduct. They have been added to the gitignore.