Call a test method by referencing its name
This change uses message methods. Instead of invoking the method
directly we use `TObject.DispatchStr` and a special string that acts as
a key to indirectly call the method. For convenience the string is the
same as the method name. It is a requirement that a message method have
only one reference variable. In this case the variable names are all
prefixed with `_` to show that we are ignoring the variable.

Because the message methods take `var` arguments they need to be
variable identifiers and not string literals.

Message methods are commonly used to allow callbacks for GUI elements
e.g. `OnClick` handlers.

Each method also has a `Writeln` added to make sure they are all still
being run. `DispatchStr` will fail silently if a string key is wrong and
no default is set.

Doing things this way is more cumbersome. It is intended that a class
should be able to run its own test methods and this should be a step
toward that goal.

Reference: https://www.freepascal.org/docs-html/current/ref/refsu31.html#x84-1080006.6.7