Test class can run test methods by using their names
This hides the details of running test methods in a `run` method. It's
simpler than the previous change but is still not as effective as
invoking the method directly. However this keeps the concept of running
a test in one place and should be closer to the goal of having an
instance run its own tests.