Test class can run all of its test methods
The class `.run` method can now be used to run all of the instance
method tests. This does require some manual book keeping: the developer
needs to record the name of the test in the `tests` array in the
`initialization` section of the unit. These are the names of the tests
that `.run` will iterate through and execute.

If a name is placed in `tests` which does not have a matching method
then `.run` will end up doing nothing without any visible failure. This
is not desirable. The name is likely a spelling mistake of a real test
method which is not getting executed. This default behavior can be
altered by overriding `TObject.DefaultHandlerStr`.