Move test running into a super class
By moving the `run` methods into a super class any class an inherit
from `TTestCase` and become a test class. The only other requirement
of a test class is that it override the `tests` class function and
record all the test method names in an array returned by the function.

I changed from a variable to a function for `tests` because it was
getting complex trying to make it into some class variable or property
that could be overridden by inheriting classes. Using a function instead
is the simplest thing that works.