Try running all existing tests using hooks
This is as far as I could take this idea. I haven't been able to get
this step to work.

Object Pascal does not seem to have the reflection required to find all
descendants of TestCase. In theory it would be possible to find all
overridden methods by using the TestCase virtual method table. In
practice it is undocumented and too difficult to decipher from the
source code. It looks like the only possible way is to use pointer
arithmetic but I can't get that to work.

I'm making a commit here to save this progress. I will try a
fundamentally different approach probably not reliant on OOP.