Simpler implementation without classes
This is a drastically simpler approach. `PascalTest` is now just a
collection of assertion procedures that all wrap the basic assert.

`ExampleTestCase` is a unit full of unit tests. It uses the assertions
that `PascalTest` provides. It also defines a `run()` procedure that
just calls all of the unit test procedures in the unit.

`TestRunner` runs all the unit tests in the repository. In this case it
uses the `run()` procedure that `ExampleTestCase` defines.

This is far from a complete testing library. It requires a little manual
effort in running all of the tests. It is however far simpler than the
object oriented approach and it works. This is enough to start using to
test other projects. That can in turn drive further development.